feat: Added weekly and monthly options
