+++ src/components/forms/RideForm.tsx
@@ -8,32 +8,9 @@ import { getNow } from "../../../shared/utils";
 import { Preferences } from "../../types";
 import { Button } from "../Button";
 import { CancelButton } from "../Button/CancelButton";
+import { RepeatingRideForm, FormValues } from "./RepeatingRideForm";
 
 const today = getNow().split("T")[0] || "";
-
-export type FormValues = {
-  id?: string;
-  name: string;
-  date: string;
-  time: string;
-  group?: string;
-  destination?: string;
-  meetPoint?: string;
-  notes?: string;
-  distance: number;
-  leader: string;
-  route: string;
-  limit?: number;
-  // Repeating ride
-  interval?: number;
-  freq?: number;
-  startDate?: string;
-  winterStartTime?: string;
-  byweekday?: number;
-  bymonth?: number;
-  bymonthday?: number;
-  byyearday?: number;
-};
 
 type RideFormProps = {
   defaultValues: FormValues;
@@ -43,6 +20,8 @@ type RideFormProps = {
   waiting: boolean;
   preferences: Preferences;
   isAdmin: boolean;
+  repeatFreq?: number;
+  selectedMonth?: number;
 };
 
 export const RideForm = ({
@@ -53,6 +32,8 @@ export const RideForm = ({
   waiting,
   preferences,
   isAdmin,
+  repeatFreq,
+  selectedMonth,
 }: RideFormProps) => {
   const [repeats, setRepeats] = useState<boolean>(true);
   const switchClass = clsx(
@@ -63,7 +44,6 @@ export const RideForm = ({
     "inline-block h-4 w-4 transform rounded-full bg-white transition",
     repeats ? "translate-x-6" : "translate-x-1"
   );
-
   const handleRepeatsChange = () => setRepeats(!repeats);
 
   return (
@@ -255,7 +235,6 @@ export const RideForm = ({
         </label>
       </div>
 
-      {/* Repeating ride section */}
       {isAdmin && (
         <div className="flex flex-row">
           <div className="pr-8">This ride repeats</div>
@@ -269,151 +248,16 @@ export const RideForm = ({
           </Switch>
         </div>
       )}
+
       {isAdmin && repeats && (
-        <>
-          <div className="grid grid-cols-3 gap-4">
-            <div className="form-control w-full">
-              <label htmlFor="interval">Every</label>
-              <select
-                id="interval"
-                className="select text-lg font-normal"
-                defaultValue={defaultValues.interval}
-                {...register("interval")}
-              >
-                <option value="1">1</option>
-                <option value="2">2</option>
-                <option value="3">3</option>
-                <option value="4">4</option>
-              </select>
-            </div>
-
-            <div className="col-span-2 form-control w-full">
-              <label htmlFor="freq">Frequency</label>
-              <select
-                id="freq"
-                aria-label="Repeating schedule freqency"
-                className="select text-lg font-normal"
-                defaultValue={defaultValues.freq}
-                {...register("freq")}
-              >
-                <option value="2">Weeks</option>
-                <option value="1">Months</option>
-                <option value="0">Years</option>
-              </select>
-            </div>
-          </div>
-          {/* TODO: display relevant list for freq */}
-          <div className="flex">
-            <div className="col-span-2 form-control w-full">
-              <label htmlFor="byweekday">On</label>
-              <select
-                id="byweekday"
-                aria-label="Repeating schedule freqency"
-                className="select text-lg font-normal"
-                defaultValue={defaultValues.byweekday}
-                {...register("byweekday")}
-              >
-                <option value="0">Monday</option>
-                <option value="1">Tuesday</option>
-                <option value="2">Wednesday</option>
-                <option value="3">Thursday</option>
-                <option value="4">Friday</option>
-                <option value="5">Saturday</option>
-                <option value="6">Sunday</option>
-              </select>
-            </div>
-          </div>
-          {/* Month */}
-          <div className="grid grid-cols-3 gap-4">
-            <div className="form-control w-full">
-              <label htmlFor="bymonthday">Day</label>
-              <select
-                id="bymonthday"
-                className="select text-lg font-normal"
-                defaultValue={defaultValues.bymonthday}
-                {...register("bymonthday")}
-              >
-                {/* TODO: programatically set days in selected month */}
-                <option value="1">1</option>
-                <option value="2">2</option>
-                <option value="3">3</option>
-                <option value="4">4</option>
-                <option value="5">5</option>
-                <option value="6">6</option>
-                <option value="7">7</option>
-                <option value="8">8</option>
-                <option value="9">9</option>
-                <option value="10">10</option>
-                <option value="11">11</option>
-                <option value="12">12</option>
-                <option value="13">13</option>
-                <option value="14">14</option>
-                <option value="15">15</option>
-              </select>
-            </div>
-
-            <div className="col-span-2 form-control w-full">
-              <label htmlFor="bymonth">Month</label>
-              <select
-                id="bymonth"
-                aria-label="Repeating schedule freqency"
-                className="select text-lg font-normal"
-                defaultValue={defaultValues.bymonth}
-                {...register("bymonth")}
-              >
-                <option value="0">January</option>
-                <option value="1">Febraury</option>
-                <option value="2">March</option>
-                <option value="3">April</option>
-                <option value="4">May</option>
-                <option value="5">June</option>
-                <option value="6">July</option>
-                <option value="7">August</option>
-                <option value="8">September</option>
-                <option value="9">October</option>
-                <option value="10">November</option>
-                <option value="11">December</option>
-              </select>
-            </div>
-          </div>
-          {/* Date and time */}
-          {/* TODO: Cannot be in past */}
-          <div className="grid grid-cols-2 gap-4">
-            <div className="flex flex-col gap-4 md:gap-8">
-              <label htmlFor="startDate" className="flex flex-col gap-1">
-                Start Date
-                <input
-                  id="startDate"
-                  type="date"
-                  min={today}
-                  className="input"
-                  defaultValue={defaultValues.startDate}
-                  {...register("startDate", {
-                    required: true,
-                  })}
-                />
-                {errors.date && (
-                  <span className="font-normal text-red-500">
-                    Start date is required
-                  </span>
-                )}
-              </label>
-            </div>
-
-            <div className="flex flex-col gap-4 md:gap-8">
-              <label htmlFor="winterStartTime" className="flex flex-col gap-1">
-                Winter Start
-                <input
-                  id="winterStartTime"
-                  type="time"
-                  className="input"
-                  defaultValue={defaultValues.winterStartTime}
-                  {...register("winterStartTime")}
-                />
-              </label>
-            </div>
-          </div>
-        </>
+        <RepeatingRideForm
+          defaultValues={defaultValues}
+          register={register}
+          errors={errors}
+          repeats={repeats}
+          repeatFreq={repeatFreq}
+          selectedMonth={selectedMonth}
+        />
       )}
 
       <div className="grid w-full grid-cols-2 gap-4 md:gap-8">
@@ -425,3 +269,5 @@ export const RideForm = ({
     </form>
   );
 };
+
+export { type FormValues };
